리팩토링: 랜딩 레이아웃 분리
