+++ src/widgets/index.ts
@@ -2,5 +2,6 @@ export * from "./canvas";
 export * from "./editor-layout";
 export * from "./editor-toolbar";
 export * from "./input-panel";
+export * from "./landing-layout";
 export * from "./layout";
 export * from "./output-panel";
