+++ src/widgets/landing-layout/ui/index.ts
@@ -1,0 +1,1 @@
+export * from "./LandingLayout";
